+++ MapleServer2/Types/DebugPrintSettings.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MapleServer2.Types;
+﻿namespace MapleServer2.Types;
 
 public class DebugPrintSettings
 {
