--- MapleWebServer/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-{
-  "profiles": {
-    "MapleWebServer": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "https://localhost:61083;http://localhost:61084"
-    }
-  }
-}
